+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `MainCategories` component.
 - `ProductDescription` content loader.
+- `Impersonate` component to  `Header`.
 
 ## [1.12.0] - 2018-08-02
 ### Added