+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Impersonate` component to  `Header`.
 
 ## [1.12.3] - 2018-08-09
 ### Fixed
@@ -22,7 +24,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `MainCategories` component.
 - `ProductDescription` content loader.
-- `Impersonate` component to  `Header`.
 
 ## [1.12.0] - 2018-08-02
 ### Added