+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Undefined behavior of the `CategoriesHightlights` schema.
+
 ## [1.12.5] - 2018-08-13
 ### Added
 - `ImpersonateCustomer` component to  `Header`.